requirements.txt
flask >= 2.1.0
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp             >= 1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc  >= 1.12.0rc2
opentelemetry-exporter-otlp-proto-http  >= 1.12.0rc2

opentelemetry-instrumentation-wsgi == 0.32b0
opentelemetry-util-http            == 0.32b0
